requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.12
aiosignal==1.3.2
APScheduler==3.11.0
attrs==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
propcache==0.2.1
python-dotenv==1.0.1
pytz==2025.1
requests==2.32.3
tzlocal==5.2
urllib3==2.3.0
yarl==1.18.3
